refactor SignUpConfirm 컴포넌트 js to ts
