+++ src/pages/SignUpConfirm.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState, useRef } from 'react';
+import { useEffect, useState, useRef } from 'react';
 import { useNavigate } from 'react-router-dom';
 import SignLayout from '../components/SignLayout';
 import fetchAchievement from '@/apis/fetchAchievement';
@@ -9,12 +9,12 @@ import { useQueryClient } from '@tanstack/react-query';
 const SignUpConfirm = () => {
   const navigate = useNavigate();
   const { setUserInfo } = userStore();
-  const baekjoonData = useRef();
-  const [userId, setUserId] = useState(null);
+  const baekjoonData = useRef<string | null>(null);
+  const [userId, setUserId] = useState<string>();
   const queryClient = useQueryClient();
 
   useEffect(() => {
-    let timeoutId;
+    let timeoutId: NodeJS.Timeout;
     const getSession = async () => {
       const session = await supabaseClient.auth.getSession();
       if (session.data.session?.user.id) {
@@ -35,22 +35,23 @@ const SignUpConfirm = () => {
           if (!baekjoonData.current) {
             // setTimeout으로 한 이유는 onAuthStateChange와 데드락을 피하기 위해 사용
             timeoutId = setTimeout(async () => {
-              let { data: baekjoon, error } = await supabaseClient
+              const { data: baekjoon } = await supabaseClient
                 .from('baekjoon')
                 .select('solved_recent')
                 .eq('id', session.user.user_metadata.baekjoon_id);
-              baekjoonData.current = baekjoon[0].solved_recent;
+              baekjoonData.current = baekjoon?.[0].solved_recent;
               // baekjoon 테이블에 있는 solved_recent 컬럼을 조회 후 없으면 크롤링
               // session에 baekjoon_id가 존재해야함
               if (
                 session.user.user_metadata.baekjoon_id &&
-                !baekjoon[0].solved_recent
+                !baekjoon?.[0].solved_recent
               ) {
-                const crawling = await queryClient.fetchQuery({
+                const res = await queryClient.fetchQuery({
                   queryKey: ['solved', session.user.user_metadata.baekjoon_id],
                   queryFn: () =>
                     fetchAchievement(session.user.user_metadata.baekjoon_id),
                 });
+                const crawling = res.data[0];
                 // 크롤링한 데이터를 가지고 baekjoon 테이블 업데이트
                 await supabaseClient
                   .from('baekjoon')
